1. Output data compression in the clusterfinder component is removed because it doesn't help (for a moment the output blocks are not filtered by the framework and all the output is sent to the tracker )
To compress the output, the AliHLTTPCCAInputDataCompressorComponent should run between the clusterfinder and the tracker. 

2. compilation warnings fixed


M    TPCLib/tracking-ca/AliHLTTPCCAInputDataCompressorComponent.cxx
M    TPCLib/tracking-ca/AliHLTTPCCATrackerComponent.cxx
M    TPCLib/AliHLTTPCClusterFinderComponent.cxx

+++ GPU/TPCCAGPUTracking/tracking-ca/AliHLTTPCCAInputDataCompressorComponent.cxx
@@ -265,7 +265,7 @@ int AliHLTTPCCAInputDataCompressorComponent::Compress( AliHLTTPCClusterData* inp
       outRow = (AliHLTTPCCACompressedClusterRow*) outCluster;
       outCluster = outRow->fClusters;  
       outputSize+= ( ( AliHLTUInt8_t * )outCluster ) -  (( AliHLTUInt8_t * )outRow);
-      if ( outputSize > (int) maxBufferSize ){
+      if ( outputSize >  maxBufferSize ){
 	ret = -ENOSPC;
  	outputSize=0;
 	break;
@@ -295,7 +295,7 @@ int AliHLTTPCCAInputDataCompressorComponent::Compress( AliHLTTPCClusterData* inp
       UInt_t iz = ( (UInt_t) z )&0x00FFFFFF;
       
       outputSize+= sizeof( AliHLTTPCCACompressedCluster );
-      if ( outputSize > (int) maxBufferSize ){
+      if ( outputSize > maxBufferSize ){
 	outputSize = 0;
 	ret = -ENOSPC;
 	break;
